refactor proposal memo replay evidence router
